+++ packages/atmosphere/src/StarsMaterial.ts
@@ -35,8 +35,6 @@ declare module 'three' {
 export interface StarsMaterialParameters
   extends AtmosphereMaterialBaseParameters {
   pointSize?: number
-  /** @deprecated Use intensity instead. */
-  radianceScale?: number
   intensity?: number
   background?: boolean
 }
@@ -66,7 +64,7 @@ export class StarsMaterial extends AtmosphereMaterialBase {
   pointSize: number
 
   constructor(params?: StarsMaterialParameters) {
-    const { pointSize, radianceScale, intensity, background, ...others } = {
+    const { pointSize, intensity, background, ...others } = {
       ...starsMaterialParametersDefaults,
       ...params
     }
@@ -90,7 +88,7 @@ export class StarsMaterial extends AtmosphereMaterialBase {
         cameraFar: new Uniform(0),
         pointSize: new Uniform(0),
         magnitudeRange: new Uniform(new Vector2(-2, 8)),
-        intensity: new Uniform(radianceScale ?? intensity),
+        intensity: new Uniform(intensity),
         ...others.uniforms
       } satisfies StarsMaterialUniforms,
       defines: {
@@ -127,16 +125,6 @@ export class StarsMaterial extends AtmosphereMaterialBase {
 
   get magnitudeRange(): Vector2 {
     return this.uniforms.magnitudeRange.value
-  }
-
-  /** @deprecated Use intensity instead. */
-  get radianceScale(): number {
-    return this.intensity
-  }
-
-  /** @deprecated Use intensity instead. */
-  set radianceScale(value: number) {
-    this.intensity = value
   }
 
   get intensity(): number {