+++ packages/atmosphere/src/r3f/Stars.tsx
@@ -35,8 +35,6 @@ export interface StarsProps
     AtmosphereMaterialProps {
   data?: ArrayBuffer | string
   pointSize?: number
-  /** @deprecated Use intensity instead. */
-  radianceScale?: number
   intensity?: number
   background?: boolean
 }
@@ -51,7 +49,7 @@ export const Stars: FC<StarsProps> = ({
 
   const [
     atmosphereParameters,
-    { pointSize, radianceScale, intensity, background, ...others }
+    { pointSize, intensity, background, ...others }
   ] = separateProps({
     ...starsMaterialParametersDefaults,
     ...contextProps,
@@ -117,7 +115,6 @@ export const Stars: FC<StarsProps> = ({
         object={material}
         {...atmosphereParameters}
         pointSize={pointSize}
-        radianceScale={radianceScale}
         intensity={intensity}
         background={background}
         depthTest={true}